requirements.txt
streamlit>=1.44.0
pandas>=2.2.0
openai>=1.0.0
python-dotenv>=1.0.1
